+++ noodles-bcf/examples/bcf_write.rs
@@ -30,7 +30,7 @@ fn main() -> io::Result<()> {
 
     let record = vcf::variant::RecordBuf::builder()
         .set_reference_sequence_name("sq0")
-        .set_position(Position::MIN)
+        .set_variant_start(Position::MIN)
         .set_reference_bases("A")
         .build();